Refactor in Resolver class
