+++ app/src/main/java/ayds/newyork/songinfo/moredetails/data/external/nytimes/artistinfo/NYTimesToArtistInfoResolver.kt
@@ -1,14 +1,22 @@
 package ayds.newyork.songinfo.moredetails.data.external.nytimes.artistinfo
 
+import ayds.newyork.songinfo.moredetails.domain.entities.ArtistInfo
 import com.google.gson.Gson
 import com.google.gson.JsonElement
 import com.google.gson.JsonObject
 import retrofit2.Response
+import java.io.IOException
 
 private const val P_RESPONSE = "response"
 private const val NO_RESULTS = "No Results"
-interface NYTimesToArtistInfoResolver{
-    fun updateInfoArtist(callResponse: Response<String>,nameArtist: String?): String
+private const val DOCS = "docs"
+private const val ABSTRACT = "abstract"
+private const val WEB_URL = "web_url" //no debería tener el valor de una url?
+
+interface NYTimesToArtistInfoResolver {
+    fun updateInfoArtist(callResponse: Response<String>, nameArtist: String?): String
+
+    fun getArtistInfoFromExternalData(serviceData: String?): ArtistInfo?
 }
 
 internal class JsonToArtistInfoResolver : NYTimesToArtistInfoResolver{
@@ -17,6 +25,38 @@ internal class JsonToArtistInfoResolver : NYTimesToArtistInfoResolver{
         val abstract = getAsJsonObject(response)
         return abstract?.asString?.replace("\\n", "\n") ?: NO_RESULTS
     }
+
+    override fun getArtistInfoFromExternalData(serviceData: String?): ArtistInfo? {
+        return try {
+            serviceData?.getFirstItem()?.let { item ->
+                ArtistInfo(
+                    item.getArtist(), item.getAbstract(), item.getUrl()
+                )
+            }
+        } catch (e: IOException) {
+            e.printStackTrace()
+            null
+        }
+    }
+
+    private fun JsonObject.getArtist(): String {
+    }
+
+    private fun JsonObject.getAbstract(): String {
+
+    }
+
+    private fun JsonObject.getUrl(): String {
+
+    }
+
+    private fun String?.getFirstItem(): JsonObject {
+        val jobj = Gson().fromJson(this, JsonObject::class.java)
+        val response = jobj.get(P_RESPONSE).asJsonObject
+        val docs = response[DOCS].asJsonArray
+        return docs[0].asJsonObject
+    }
+
     private fun getJson(callResponse: Response<String>): JsonObject {
         val gson = Gson()
         return gson.fromJson(callResponse.body(), JsonObject::class.java)
@@ -26,4 +66,23 @@ internal class JsonToArtistInfoResolver : NYTimesToArtistInfoResolver{
         return response["docs"].asJsonArray[0].asJsonObject["abstract"]
     }
 
+    private fun getInfoFromDataBase(): ArtistInfo? {
+        return if (artistName != null) NYTimesArtistInfoLocalStorageImpl.getArtistInfo(artistName!!) else null
+    }
+
+    private fun jsonToArtistInfo(jobj: JsonObject?): ArtistInfo? {
+        if (jobj == null)
+            return null
+        val response = jobj.get(P_RESPONSE).asJsonObject
+        val docs = response[DOCS].asJsonArray
+        val abstract = if (docs.size() == 0) getTextFromAbstract(null) else getTextFromAbstract(
+            docs.get(0).asJsonObject.get(ABSTRACT).asString
+        )
+        val url = if (docs.size() == 0) null else docs.get(0).asJsonObject.get(WEB_URL).asString
+        return artistName?.let {
+            ArtistInfo(
+                it, abstract, url
+            )
+        }
+    }
 }
